Plugin pom updated to use hudson-plugin-parent 2.0.0. Tests were refactored.
