+++ pom.xml
@@ -22,45 +22,44 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 THE SOFTWARE.
 -->
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-  <parent>
-    <groupId>org.jvnet.hudson.plugins</groupId>
-    <artifactId>plugin</artifactId>
-    <version>1.388</version>
-  </parent>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+    <parent>
+        <groupId>org.jvnet.hudson.plugins</groupId>
+        <artifactId>hudson-plugin-parent</artifactId>
+        <version>2.0.0</version>
+    </parent>
 
-  <artifactId>subversion</artifactId>
-  <packaging>hpi</packaging>
-  <name>Hudson Subversion Plug-in</name>
-  <version>1.24-SNAPSHOT</version>
-  <url>http://wiki.hudson-ci.org/display/HUDSON/Subversion+Plugin</url>
-  <licenses>
-    <license>
-      <name>MIT license</name>
-      <comments>All source code is under the MIT license.</comments>
-    </license>
-  </licenses>
+    <artifactId>subversion</artifactId>
+    <packaging>hpi</packaging>
+    <name>Hudson Subversion Plug-in</name>
+    <version>2.0.0</version>
+    <url>http://wiki.hudson-ci.org/display/HUDSON/Subversion+Plugin</url>
+    <licenses>
+        <license>
+            <name>MIT license</name>
+            <comments>All source code is under the MIT license.</comments>
+        </license>
+    </licenses>
 
-  <dependencies>
-    <dependency>
-      <groupId>org.jvnet.hudson.svnkit</groupId>
-      <artifactId>svnkit</artifactId>
-      <version>1.3.4-hudson-2</version>
-    </dependency>
-  </dependencies>
+    <dependencies>
+        <dependency>
+            <groupId>org.jvnet.hudson.svnkit</groupId>
+            <artifactId>svnkit</artifactId>
+            <version>1.3.4-hudson-2</version>
+        </dependency>
+        <dependency>
+            <groupId>org.jvnet.hudson.main</groupId>
+            <artifactId>hudson-test-harness</artifactId>
+            <version>2.0.0</version>
+        </dependency>
+    </dependencies>
 
-  <developers>
-    <developer>
-      <name>Many</name>
-      <id>kohsuke abayer dodok1 dty huybrechts mindless pgweiss stephenconnolly rseguy etc</id>
-    </developer>
-  </developers>
-
-  <distributionManagement>
-    <repository>
-      <id>java.net-m2-repository</id>
-      <url>http://maven.hudson-labs.org:8081/content/repositories/releases/</url>
-    </repository>
-  </distributionManagement>
+    <developers>
+        <developer>
+            <name>Many</name>
+            <id>kohsuke abayer dodok1 dty huybrechts mindless pgweiss stephenconnolly rseguy etc</id>
+        </developer>
+    </developers>
 </project>
