+++ src/test/java/hudson/scm/AbstractSubversionTest.java
@@ -11,6 +11,8 @@ import org.jvnet.hudson.test.HudsonTestCase;
 import java.io.File;
 import java.io.IOException;
 import java.net.URL;
+import org.tmatesoft.svn.core.auth.ISVNAuthenticationManager;
+import org.tmatesoft.svn.core.wc.SVNWCUtil;
 
 /**
  * Base class for Subversion related tests.
@@ -19,6 +21,8 @@ import java.net.URL;
  */
 public abstract class AbstractSubversionTest extends HudsonTestCase  {
     protected DescriptorImpl descriptor;
+    protected String kind = ISVNAuthenticationManager.PASSWORD;
+
 
     @Override
     protected void setUp() throws Exception {
@@ -43,6 +47,12 @@ public abstract class AbstractSubversionTest extends HudsonTestCase  {
         }
         return launcher.launch().cmds(
                 "svnserve","-d","--foreground","-r",repo.getAbsolutePath()).pwd(repo).start();
+    }
+
+    protected ISVNAuthenticationManager createInMemoryManager() {
+        ISVNAuthenticationManager m = SVNWCUtil.createDefaultAuthenticationManager(hudson.root, null, null, false);
+        m.setAuthenticationProvider(descriptor.createAuthenticationProvider(null));
+        return m;
     }
 
     static {
