+++ src/test/java/hudson/scm/PerJobCredentialStoreTest.java
@@ -2,17 +2,36 @@ package hudson.scm;
 
 import hudson.Proc;
 import hudson.model.FreeStyleProject;
+import hudson.model.Result;
 import java.io.IOException;
 import java.io.PrintWriter;
+import org.junit.Ignore;
 import org.jvnet.hudson.test.Bug;
+import org.tmatesoft.svn.core.SVNCancelException;
+import org.tmatesoft.svn.core.SVNException;
+import org.tmatesoft.svn.core.SVNURL;
+import org.tmatesoft.svn.core.auth.ISVNAuthenticationManager;
+import org.tmatesoft.svn.core.auth.SVNPasswordAuthentication;
+import org.tmatesoft.svn.core.io.SVNRepository;
+import org.tmatesoft.svn.core.io.SVNRepositoryFactory;
+
+import static hudson.scm.SubversionSCM.compareSVNAuthentications;
 
 /**
  * @author Kohsuke Kawaguchi
  */
+@Ignore
 public class PerJobCredentialStoreTest extends AbstractSubversionTest {
     private static final String testSvnUser = "user";
     private static final String testSvnPassword = "userpass";
     private static final String testSvnRealm = "Test realm";
+    private static final String GUEST_ACCESS_REPOSITORY_RESOURCE = "guest_access_svn.zip";
+    private static final String realm = "<svn://localhost:3690>";
+    private static final String SVN_URL = "svn://localhost/bob";
+    private static final String GUEST_USER_LOGIN = "guest";
+    private static final String GUEST_USER_PASSWORD = "guestpass";
+    private static final String BOGUS_USER_LOGIN = "bogus";
+    private static final String BOGUS_USER_PASSWORD = "boguspass";
 
     private PerJobCredentialStore credentialStore;
 
@@ -40,10 +59,11 @@ public class PerJobCredentialStoreTest extends AbstractSubversionTest {
     /**
      * There was a bug that credentials stored in the remote call context was serialized wrongly.
      */
-    @Bug(8061)
+//    @Bug(8061)
     public void testRemoteBuild() throws Exception {
-        Proc p = runSvnServe(SubversionSCMTest.class.getResource("HUDSON-1379.zip"));
+        Proc p = null;
         try {
+            p = runSvnServe(getClass().getResource("HUDSON-1379.zip"));
             FreeStyleProject b = createFreeStyleProject();
             b.setScm(new SubversionSCM("svn://localhost/bob"));
             b.setAssignedNode(createSlave());
@@ -55,7 +75,110 @@ public class PerJobCredentialStoreTest extends AbstractSubversionTest {
             PerJobCredentialStore store = new PerJobCredentialStore(b);
             assertFalse(store.isEmpty());   // credential store should contain a valid entry
         } finally {
+            if(p!= null){
+                p.kill();
+            }
+        }
+    }
+
+    /**
+     * Even if the default providers remember bogus passwords, Hudson should still attempt what it knows.
+     */
+    @Bug(3936)
+    public void test3936() throws Exception {
+        //Start local svn repository
+        Proc server = runSvnServe(getClass().getResource(GUEST_ACCESS_REPOSITORY_RESOURCE));
+        SVNURL repo = SVNURL.parseURIDecoded(SVN_URL);
+        try {
+
+            // creates a purely in memory auth manager
+            ISVNAuthenticationManager m = createInMemoryManager();
+
+            // double check that it really knows nothing about the fake repo
+            try {
+                m.getFirstAuthentication(kind, realm, repo);
+                fail();
+            } catch (SVNCancelException e) {
+                // yep
+            }
+
+            // teach a bogus credential and have SVNKit store it.
+            SVNPasswordAuthentication bogus = new SVNPasswordAuthentication(BOGUS_USER_LOGIN, BOGUS_USER_PASSWORD, true);
+            m.acknowledgeAuthentication(true, kind, realm, null, bogus);
+            assertTrue(compareSVNAuthentications(m.getFirstAuthentication(kind, realm, repo), bogus));
+            try {
+                attemptAccess(repo, m);
+                fail("SVNKit shouldn't yet know how to access");
+            } catch (SVNCancelException e) {
+            }
+
+            // make sure the failure didn't clean up the cache,
+            // since what we want to test here is Hudson trying to supply its credential, despite the failed cache
+            assertTrue(compareSVNAuthentications(m.getFirstAuthentication(kind, realm, repo), bogus));
+
+            // now let Hudson have the real credential
+            // can we now access the repo?
+            descriptor.postCredential(null, repo.toDecodedString(), GUEST_USER_LOGIN, GUEST_USER_PASSWORD, null, new PrintWriter(System.out));
+            attemptAccess(repo, m);
+        } finally {
+            server.kill();
+        }
+    }
+
+    @Bug(1379)
+    public void testMultipleCredentialsPerRepo() throws Exception {
+        Proc p = runSvnServe(getClass().getResource("HUDSON-1379.zip"));
+        try {
+            FreeStyleProject b = createFreeStyleProject();
+            b.setScm(new SubversionSCM(SVN_URL));
+
+            FreeStyleProject c = createFreeStyleProject();
+            c.setScm(new SubversionSCM("svn://localhost/charlie"));
+
+            // should fail without a credential
+            assertBuildStatus(Result.FAILURE, b.scheduleBuild2(0).get());
+            descriptor.postCredential(b, SVN_URL, "bob", "bob", null, new PrintWriter(System.out));
+            buildAndAssertSuccess(b);
+
+            assertBuildStatus(Result.FAILURE, c.scheduleBuild2(0).get());
+            descriptor.postCredential(c, "svn://localhost/charlie", "charlie", "charlie", null,
+                new PrintWriter(System.out));
+            buildAndAssertSuccess(c);
+
+            // b should still build fine.
+            buildAndAssertSuccess(b);
+        } finally {
             p.kill();
         }
+    }
+
+
+    @Bug(1379)
+    public void testSuperUserForAllRepos() throws Exception {
+        Proc p = runSvnServe(getClass().getResource("HUDSON-1379.zip"));
+        try {
+            FreeStyleProject b = createFreeStyleProject();
+            b.setScm(new SubversionSCM(SVN_URL));
+
+            FreeStyleProject c = createFreeStyleProject();
+            c.setScm(new SubversionSCM("svn://localhost/charlie"));
+
+            // should fail without a credential
+            assertBuildStatus(Result.FAILURE, b.scheduleBuild2(0).get());
+            assertBuildStatus(Result.FAILURE, c.scheduleBuild2(0).get());
+
+            // but with the super user credential both should work now
+            descriptor.postCredential(b, SVN_URL, "alice", "alice", null, new PrintWriter(System.out));
+            buildAndAssertSuccess(b);
+            buildAndAssertSuccess(c);
+        } finally {
+            p.kill();
+        }
+    }
+
+    private void attemptAccess(SVNURL repo, ISVNAuthenticationManager m) throws SVNException {
+        SVNRepository repository = SVNRepositoryFactory.create(repo);
+        repository.setAuthenticationManager(m);
+        repository.testConnection();
     }
 }
